% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyzeVariants.R
\name{analyzeVariants}
\alias{analyzeVariants}
\title{Calculate and process Variants}
\usage{
analyzeVariants()
}
\value{
Nothing
}
\description{
Calculate and process Variants
}
\author{
Jens Reeder
}

