% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeVariants.R
\name{annotateVariants}
\alias{annotateVariants}
\title{Annotate variants via vep}
\usage{
annotateVariants(vcf.file)
}
\arguments{
\item{vcf.file}{A character vector pointing to a VCF (or gzipped VCF) file}
}
\value{
Path to a vcf file with variant annotations
}
\description{
Annotate variants via vep
}
\author{
Jens Reeder
}

