% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildShortReadReports}
\alias{buildShortReadReports}
\title{Build a ShortRead report}
\usage{
buildShortReadReports(save_dir, paired_ends)
}
\arguments{
\item{save_dir}{Save directory of a pipeline run}

\item{paired_ends}{A logical, indicating whether reads are paired}
}
\value{
Nothing
}
\description{
Build a ShortRead report
}
\author{
Gregoire Pau
}
\keyword{internal}

