% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeVariants.R
\name{buildTallyParam}
\alias{buildTallyParam}
\title{Build tally parameters}
\usage{
buildTallyParam()
}
\value{
a \code{VariantTallyParam} object
}
\description{
Build tally parameters
}
\author{
Gregoire Pau
}
\keyword{internal}

