% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildTP53FastaGenome}
\alias{buildTP53FastaGenome}
\title{buildTP53FastaGenome}
\usage{
buildTP53FastaGenome()
}
\value{
Path to tp53 genome directory
}
\description{
create fasta genome file of TP53 genome
}
\author{
Jens Reeder
}
\keyword{internal}

