% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{buildTP53GenomeTemplate}
\alias{buildTP53GenomeTemplate}
\title{buildTP53GenomeTemplate}
\usage{
buildTP53GenomeTemplate()
}
\value{
Path to tp53 template file
}
\description{
Create a tp53 config template
}
\author{
Jens Reeder
}
\keyword{internal}

