\name{calcConfusionMatrix}
\alias{calcConfusionMatrix}
\title{Calculate Confusion Matrix}
\usage{
  calcConfusionMatrix(variants)
}
\arguments{
  \item{variants}{Variants as GRanges}
}
\value{
  confusion matrix as table
}
\description{
  Calculate Confusion Matrix
}
\details{
  Calculate Confusion Matrix from Granges object containing
  variants
}
\author{
  Cory Barr, Jens Reeder
}

