% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coverage.R
\name{calculateCoverage}
\alias{calculateCoverage}
\title{Calculate read coverage}
\usage{
calculateCoverage()
}
\value{
Nothing
}
\description{
Calculate read coverage
}
\author{
Jens Reeder
}
\keyword{internal}

