\name{checkConfig.detectAdapterContam}
\alias{checkConfig.detectAdapterContam}
\title{Check adapter contamination configuration values}
\usage{
  checkConfig.detectAdapterContam()
}
\value{
  Nothing, but will throw error instead.
}
\description{
  Performs all configuration checks for adapter
  copntamination related values
}

