% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConfig.R
\name{checkConfig}
\alias{checkConfig}
\title{Check configuration}
\usage{
checkConfig()
}
\value{
Nothing. Individual checks will throw error instead.
}
\description{
Performs all configuration checks
}
\keyword{internal}

