% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkGATKJar}
\alias{checkGATKJar}
\title{Check for the GATK jar file}
\usage{
checkGATKJar(path = getOption("gatk.path"))
}
\arguments{
\item{path}{Path to the GATK jar file}
}
\value{
TRUE if tool can be called, FALSE otherwise
}
\description{
Check for the GATK jar file
}

