% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeBamStats}
\alias{computeBamStats}
\title{Compute record statistics from a bam file}
\usage{
computeBamStats(bam)
}
\arguments{
\item{bam}{A character string containing an existing bam file}
}
\value{
A numeric vector
}
\description{
Compute record statistics from a bam file
}
\details{
The statistics are additive over chunks/lanes.
}
\author{
Gregoire Pau
}
\keyword{internal}

