% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countGenomicFeatures.R
\name{countGenomicFeatures}
\alias{countGenomicFeatures}
\title{Count overlaps with genomic features}
\usage{
countGenomicFeatures()
}
\value{
Nothing
}
\description{
Count overlaps with genomic features
}
\author{
Gegoire Pau
}
\keyword{internal}

