% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countGenomicFeatures.R
\name{countGenomicFeaturesChunk}
\alias{countGenomicFeaturesChunk}
\title{Count reads by genomic Feature}
\usage{
countGenomicFeaturesChunk(save_dir, genomic_features)
}
\arguments{
\item{save_dir}{PAth to a pipeline run's save dir}

\item{genomic_features}{A list of genomic features to tally}
}
\value{
Nothing
}
\description{
Count reads by genomic Feature
}
\details{
given a BAM-file output from gsnap (with the MD tag), count
hits to exons, genes, ncRNAs, etc.
and quantify miRNA/ncRNA contaminatino
}
\author{
Cory Barr
}
\keyword{internal}

