% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectRRNA.R
\name{detectRRNA}
\alias{detectRRNA}
\title{Detect rRNA Contamination in Reads}
\usage{
detectRRNA(lreads, remove_tmp_dir = TRUE, save_dir = NULL)
}
\arguments{
\item{lreads}{A list of ShortReadQ objects}

\item{remove_tmp_dir}{boolean indicating whether or not to delete temp directory of gsnap results}

\item{save_dir}{Save directory}
}
\value{
a named logical vector indicating if a read has rRNA contamination
}
\description{
Returns a named vector indicating if a read ID has rRNA contamination or not
}
\details{
Given a genome and fastq data, each read in the fastq data is
aligned against the rRNA sequences for that genome
}
\author{
Cory Barr
}

