% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FastQStreamer.release}
\alias{FastQStreamer.release}
\title{Close the FastQStreamer}
\usage{
FastQStreamer.release()
}
\value{
Nothing
}
\description{
Close the FastQStreamer
}
\author{
Gregoire Pau
}
\seealso{
FastQStreamer.init
}
\keyword{internal}

