% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filterQuality}
\alias{filterQuality}
\title{Filter reads by quality}
\usage{
filterQuality(lreads)
}
\arguments{
\item{lreads}{A list of ShortReadQ objects}
}
\value{
A list of quality filterered ShortReadQ objects
}
\description{
Filtering reads by quality score.
Discards reads that have more than
a fraction of X nucleotides with a score below Y.
}
\details{
X and Y are controlled by global config variables
X: filterQuality.minFrac
Y: filterQuality.minQuality
}
\keyword{internal}

