% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGenomicFeaturesFromTxDb.R
\name{generateSingleGeneDERs}
\alias{generateSingleGeneDERs}
\title{generateSingleGeneDERs}
\usage{
generateSingleGeneDERs(txdb)
}
\arguments{
\item{txdb}{A transcript DB object}
}
\value{
single gene DERs
}
\description{
Generate DEXSeq-ready exons
}
\details{
generateSingleGeneDERs() generates exons by:
1) disjoining the whole exon set
2) keeping only the exons of coding regions
3) keeping only the exons that belong to unique genes
}

