% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAdapterSeqs}
\alias{getAdapterSeqs}
\title{Read list of Illumina adapter seqs from package data}
\usage{
getAdapterSeqs(paired_ends, force_paired_end_adapter, pair_num = 1)
}
\arguments{
\item{paired_ends}{Dow we have paired ends reads?}

\item{force_paired_end_adapter}{Force paired end adapters for single end reads?}

\item{pair_num}{1 for forward read, 2 for reverse read}
}
\value{
The adapter seq as string
}
\description{
Read list of Illumina adapter seqs from package data
}
\keyword{internal}

