% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getChunkDirs}
\alias{getChunkDirs}
\title{Get the list of chunk directories}
\usage{
getChunkDirs()
}
\value{
List of chunk directories
}
\description{
Get the list of chunk directories
}
\author{
Gregoire Pau
}
\keyword{internal}

