% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{getConfig.logical}
\alias{getConfig.logical}
\title{Check if a config parameter has a logical value}
\usage{
getConfig.logical(p, ...)
}
\arguments{
\item{p}{Name of parameter}

\item{...}{extra params passed to getConfig}
}
\value{
Logical value of parameter
}
\description{
Throws exception if value is not logical
}
\keyword{internal}

