% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\name{getConfig.vector}
\alias{getConfig.vector}
\title{Return values of a config variable as vector}
\usage{
getConfig.vector(p, ...)
}
\arguments{
\item{p}{Name of parameter}

\item{...}{extra params passed to getConfig}
}
\value{
value of config param as vector
}
\description{
Return values of a config variable as vector
}
\keyword{internal}

