% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{getNumericVectorDataFromFile}
\alias{getNumericVectorDataFromFile}
\title{Load data as numerical values}
\usage{
getNumericVectorDataFromFile(dir_path, object_name)
}
\arguments{
\item{dir_path}{Save dir of a pipeline run}

\item{object_name}{Object name}
}
\value{
loaded data as table of numbers
}
\description{
Load data as numerical values
}
\author{
Jens Reeder
}
\keyword{internal}

