\name{getRDataFromFile}
\alias{getRDataFromFile}
\title{Load a saved RData file from a pipeline run}
\usage{
  getRDataFromFile(dirPath, object_name)
}
\arguments{
  \item{dirPath}{Save dir of a pipeline run}

  \item{object_name}{object name}
}
\value{
  loaded RData object
}
\description{
  Load a saved RData file from a pipeline run
}
\keyword{internal}

