% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTabDataFromFile}
\alias{getTabDataFromFile}
\title{Load tabular data from the NGS pipeline result directory}
\usage{
getTabDataFromFile(save_dir, object_name)
}
\arguments{
\item{save_dir}{A character string containing an NGS pipeline output directory.}

\item{object_name}{A character string ontaining the regular expression matching a filename in dir_path}
}
\value{
A data frame.
}
\description{
Load tabular data from the NGS pipeline result directory
}

