% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getTraceback}
\alias{getTraceback}
\title{Get traceback from tryKeepTraceback()}
\usage{
getTraceback(mto)
}
\arguments{
\item{mto}{An object of the try-error class}
}
\value{
Traceback as a string
}
\description{
Get traceback from tryKeepTraceback()
}
\keyword{internal}

