\name{getTransitionRatio}
\alias{getTransitionRatio}
\title{Get Nucleotide Transition Ratio}
\usage{
  getTransitionRatio(x, y)
}
\arguments{
  \item{x}{DNAStringSet}

  \item{y}{DNAStringSet}
}
\value{
  numeric vector of length 1
}
\description{
  Get Nucleotide Transition Ratio
}
\details{
  Get Nucleotide Transition Ratio
}
\author{
  Cory Barr
}

