% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{hashCoverage}
\alias{hashCoverage}
\title{Hashing function for coverage}
\usage{
hashCoverage(cov)
}
\arguments{
\item{cov}{A SimpleRleList object}
}
\value{
A numeric
}
\description{
Hashing function for coverage
}
\author{
Gregoire Pau
}

