% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{hashVariants}
\alias{hashVariants}
\title{Hashing function for variants}
\usage{
hashVariants(var)
}
\arguments{
\item{var}{A GRanges object}
}
\value{
A numeric
}
\description{
Hashing function for variants
}
\author{
Gregoire Pau
}

