% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{hashVector}
\alias{hashVector}
\title{Hashing function for vector}
\usage{
hashVector(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A numeric
}
\description{
Hashing function for vector
}
\author{
Gregoire Pau
}

