% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPipeline.R
\name{initDirs}
\alias{initDirs}
\title{Set up NGS output dir}
\usage{
initDirs()
}
\value{
Nothing
}
\description{
Set up NGS output dir (using save_dir from getConfig)
}
\author{
Gregoire Pau
}
\keyword{internal}

