% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initPipeline.R
\name{initLogger}
\alias{initLogger}
\title{Init loggers}
\usage{
initLogger()
}
\value{
Nothing
}
\description{
Init loggers (output dir log, using save_dir from getConfig, and console log)
}
\author{
Gregoire Pau
}
\keyword{internal}

