\name{isConfig}
\alias{isConfig}
\title{Test the presence of the parameter in the current config}
\usage{
  isConfig(parameter)
}
\arguments{
  \item{parameter}{Name of parameter}
}
\value{
  TRUE if present, FALSE otherwise
}
\description{
  Test the presence of the parameter in the current config
}
\keyword{internal}

