% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{listIterator.init}
\alias{listIterator.init}
\title{Create a iterator on a list}
\usage{
listIterator.init(x)
}
\arguments{
\item{x}{A list.}
}
\value{
Nothing
}
\description{
Create a iterator on a list
}
\details{
Only one listIterator object can be open at any time.
}
\author{
Gregoire Pau
}
\keyword{internal}

