% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logdebug}
\alias{logdebug}
\title{Log debug using the logging package}
\usage{
logdebug(msg)
}
\arguments{
\item{...}{Arguments passed to logging::logdebug}
}
\value{
Nothing
}
\description{
Log debug (with a try statement)
}
\author{
Gregoire Pau
}
\keyword{internal}

