% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{loginfo}
\alias{loginfo}
\title{Log info using the logging package}
\usage{
loginfo(msg)
}
\arguments{
\item{...}{Arguments passed to logging::loginfo}
}
\value{
Nothing
}
\description{
Log info (with a try statement)
}
\author{
Gregoire Pau
}
\keyword{internal}

