% Generated by roxygen2 (4.0.2): do not edit by hand
\name{markDups}
\alias{markDups}
\title{markDups}
\usage{
markDups()
}
\value{
Nothing
}
\description{
Mark duplicates in pipeline context
}
\details{
High level function call to mark duplicates
in the analyzed.bam file of a pipelin run.
}
\author{
Jens Reeder
}

