% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessReads.R
\name{mergePreprocessReads}
\alias{mergePreprocessReads}
\title{Merge after preprocessReads}
\usage{
mergePreprocessReads(indirs, outdir, prepend_str)
}
\arguments{
\item{indirs}{A character vector, indicating which directories have to be merged}

\item{outdir}{A character string indicating the output directory (which must exist)}

\item{prepend_str}{A character string, containing a prefix going to be appended on all output result files}
}
\value{
Nothing
}
\description{
Merge detectAdapterContam, merge preprocessed reads, create summary preprocess, build shortReadReport, remove processed
}
\author{
Gregoire Pau
}
\keyword{internal}

