% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trimReads.R
\name{mergeReads}
\alias{mergeReads}
\title{Merge paired ends reads}
\usage{
mergeReads(reads, save_dir = NULL)
}
\arguments{
\item{reads}{A set of reads as ShortReadQ object}

\item{save_dir}{Save directory}
}
\value{
A single ShortReadQ object
}
\description{
Merges overlapping paired end reads using the pear programm
via the peared package
}
\keyword{internal}

