\name{mergeSummaryAlignment}
\alias{mergeSummaryAlignment}
\title{Merge summary alignments}
\usage{
mergeSummaryAlignment(indirs, outdir, prepend_str)
}
\arguments{
  \item{indirs}{A character vector, indicating which
  directories have to be merged}

  \item{outdir}{A character string indicating the output
  directory (which must exist)}

  \item{prepend_str}{A character string, containing a
  prefix going to be appended on all output result files}
}
\value{
Nothing
}
\description{
Merge summary alignments
}
\author{
Gregoire Pau
}
\keyword{internal}

