\name{plotDepthByStrand}
\alias{plotDepthByStrand}
\title{Plot Read Depth of Variants by Strand}
\usage{
plotDepthByStrand(gr, file = NULL)
}
\arguments{
  \item{gr}{A GRanges with additional elementMetadata
  describing variants as created by HTSeqGenie}

  \item{file}{Where to write the output}
}
\value{
0
}
\description{
Plot Read Depth of Variants by Strand
}
\details{
Plot Read Depth of Variants by Strand
}
\author{
Cory Barr
}
\keyword{internal}

