\name{plotDF}
\alias{plotDF}
\title{Make continuous plots of distribution function}
\usage{
plotDF(df, ylab, xlab, filename)
}
\arguments{
  \item{df}{distribution function, given as absolute count
  and percent}

  \item{filename}{plots will be saved under [filename].png
  and [filename].pdf}

  \item{xlab}{label of x axis}

  \item{ylab}{label of y axis}
}
\value{
Nothing, creates two files instead
}
\description{
Make continuous plots of distribution function
}
\author{
Jens Reeder
}
\keyword{internal}

