\name{plotMutationMatrix}
\alias{plotMutationMatrix}
\title{Plot Mutation Matrix}
\usage{
  plotMutationMatrix(variants)
}
\arguments{
  \item{variants}{Variants as Granges computed by SNVsOmuC}
}
\value{
  Nothing, stores plots as files
}
\description{
  Plot Mutation Matrix
}
\details{
  Plot Mutation Matrix
}
\author{
  Jens Reeder, Jeremiah Degenhardt
}

