\name{plotVariantDepth}
\alias{plotVariantDepth}
\title{Plot the variant depth}
\usage{
  plotVariantDepth(variants)
}
\arguments{
  \item{variants}{Variants as computed by SVNsOmuC}
}
\value{
  Name of plot file
}
\description{
  Plot the variant depth
}
\details{
  Plot the variant depth
}
\author{
  Cory Barr, Jens Reeder
}

