% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessReads.R
\name{preprocessReadsChunk}
\alias{preprocessReadsChunk}
\title{Preprocess a chunk}
\usage{
preprocessReadsChunk(lreads, save_dir = NULL)
}
\arguments{
\item{lreads}{A list of GRanges objects, containing the reads}

\item{save_dir}{Save directory of a pipeline run}
}
\value{
save_dir Save directory of a pipeline run
}
\description{
Preprocess a chunk
}
\author{
Gregoire Pau
}
\keyword{internal}

