% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{readInputFiles}
\alias{readInputFiles}
\title{Read FastQ input files}
\usage{
readInputFiles()
}
\value{
Reads as list of ShortRead objects
}
\description{
Uses the global config to find input files
}
\keyword{internal}

