% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{readRNASeqEnds}
\alias{readRNASeqEnds}
\title{Read single/paired End Bam Files}
\usage{
readRNASeqEnds(filename, paired_ends, remove.strandness = TRUE)
}
\arguments{
\item{filename}{Path to a bam file}

\item{paired_ends}{A logical indicating whether the reads are paired}

\item{remove.strandness}{A logical indicating whether read strands should be set to "*".}
}
\value{
GRangesList
}
\description{
Read single/paired end BAM files with requested columns from the BAM
}
\author{
Cory Barr
}
\keyword{internal}

