% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gatk.R
\name{realignIndels}
\alias{realignIndels}
\title{realignIndels}
\usage{
realignIndels()
}
\value{
Nothing
}
\description{
Realign indels in pipeline context
}
\details{
High level function call to realign
indels in the analyzed.bam file using GATK
}
\author{
Jens Reeder
}

