% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gatk.R
\name{realignIndelsGATK}
\alias{realignIndelsGATK}
\title{Realign indels via GATK}
\usage{
realignIndelsGATK(bam.file)
}
\arguments{
\item{bam.file}{Path to bam.file}
}
\value{
Path to realigned bam file
}
\description{
Realing indels using the GATK tools RealignerTargetCreator
and IndelRealigner.
Requires a GATK compatible genome with a name matching
the alignment genome to be installed in 'path.gatk_genome'
}
\details{
Since GATKs IndelRealigner is not parallelized, we run it in
parallel per chromosome.
}
\author{
Jens Reeder
}

