% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{resource}
\alias{resource}
\title{Reload package source code}
\usage{
resource(dirname = ".")
}
\arguments{
\item{dirname}{Directore with files to source}
}
\value{
Nothing
}
\description{
When developing code this function can be used to quickly
reload all of the packages code, without installing it.
}
\keyword{internal}

