% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countGenomicFeatures.R
\name{rpkm}
\alias{rpkm}
\title{Calculate RPKM}
\usage{
rpkm(counts, widths, nbreads)
}
\arguments{
\item{counts}{A vector of counts}

\item{widths}{vector of the width of each bin the counts were perfomred on}

\item{nbreads}{vector containing number of reads mapped to each bin}
}
\value{
vector of RPKMs
}
\description{
Calculate RPKM
}
\author{
Gregoire Pau
}
\keyword{internal}

