\name{runPreprocessReads}
\alias{runPreprocessReads}
\title{Run the preprocesing steps of the pipeline}
\usage{
runPreprocessReads(config_filename, config_update)
}
\arguments{
  \item{config_filename}{Path to configuration file}

  \item{config_update}{List of name value pairs that will
  update the config parameters}
}
\value{
Nothing
}
\description{
Runs the preprocesing steps of the pipeline
}
\author{
Jens Reeder
}
\keyword{internal}

