\name{safe.yield}
\alias{safe.yield}
\title{Overloaded yield(...) method catching truncated exceptions for FastqStreamer}
\usage{
  safe.yield(fqs)
}
\arguments{
  \item{fqs}{An instance from the FastqSampler or
  FastqStreamer class.}
}
\value{
  Same as FastqStreamer::yield
}
\description{
  Overloaded yield(...) method catching truncated
  exceptions for FastqStreamer
}
\author{
  Gregoire Pau
}
\keyword{internal}

