% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{setChunkDir}
\alias{setChunkDir}
\title{Set the base directory for the chunks}
\usage{
setChunkDir()
}
\value{
path to chunk dir
}
\description{
Set the base directory for the chunks
}
\author{
Jens Reeder
}
\keyword{internal}

