\name{setupTestFramework}
\alias{setupTestFramework}
\title{setup test framework}
\usage{
  setupTestFramework(config.filename,
    config.update = list(), testname = "test",
    package = "HTSeqGenie")
}
\arguments{
  \item{config.filename}{configuration file}

  \item{config.update}{update list of config values}

  \item{testname}{name of test case}

  \item{package}{name of package}
}
\value{
  the created temp directory
}
\description{
  setup test framework
}
\keyword{internal}

