\name{statCountFeatures}
\alias{statCountFeatures}
\title{Compute statistics on count features}
\usage{
  statCountFeatures(save_dir, feature = "counts_gene")
}
\arguments{
  \item{save_dir}{A character string containing a NGS
  analysis directory}

  \item{feature}{A character string containing a features
  name. Default is "counts_gene".}
}
\value{
  A numeric vector containing statistics about features.
}
\description{
  Compute statistics on count features
}
\author{
  Gregoire Pau
}
\keyword{internal}

