% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trimReads}
\alias{trimReads}
\title{Trim/truncate a set of reads}
\usage{
trimReads(lreads, trim_len)
}
\arguments{
\item{lreads}{A list of ShortReadQ objects}

\item{trim_len}{The length reads will be truncated to}
}
\value{
A list of truncated ShortReadQ objects
}
\description{
Trim/truncate a set of reads
}
\keyword{internal}

