% Generated by roxygen2 (4.0.2): do not edit by hand
\name{truncateReads}
\alias{truncateReads}
\title{Trim/truncate a set of reads}
\usage{
truncateReads(reads, trim_len)
}
\arguments{
\item{reads}{A set of reads as ShortReadQ object}

\item{trim_len}{The length reads will be truncated to}
}
\value{
A truncated ShortReadQ object
}
\description{
Trim/truncate a set of reads
}
\keyword{internal}

