% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tryKeepTraceback}
\alias{tryKeepTraceback}
\title{Wrapper around try-catch}
\usage{
tryKeepTraceback(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\value{
Result of expression or error if thrown
}
\description{
Wrapper around try-catch
}
\keyword{internal}

