% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeVariants.R
\name{vcfStat}
\alias{vcfStat}
\title{Compute stats on a VCF file}
\usage{
vcfStat(vcf.filename)
}
\arguments{
\item{vcf.filename}{A character pointing to a VCF (or gzipped VCF) file}
}
\value{
A numeric vector
}
\description{
Compute stats on a VCF file
}
\author{
Gregoire Pau
}

