% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wrap.callVariants}
\alias{wrap.callVariants}
\title{Variant calling}
\usage{
wrap.callVariants(bam.file)
}
\arguments{
\item{bam.file}{Aligned reads as bam file}
}
\value{
Variants as Vranges
}
\description{
Call Variants in the pipeline framework
}
\details{
A wrapper around VariantTools callVariant
framework.
}
\author{
Jens Reeder
}

