% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeConfig}
\alias{writeConfig}
\title{Write a config file}
\usage{
writeConfig(config.filename)
}
\arguments{
\item{config.filename}{Optional name of output file. If missing, print the config file on the standard output.}
}
\value{
Name of saved file
}
\description{
Writes the currently active configuration to file
}
\keyword{internal}

