% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{writeFastQFiles}
\alias{writeFastQFiles}
\title{Write reads to file}
\usage{
writeFastQFiles(lreads, dir, filename1, filename2)
}
\arguments{
\item{lreads}{List of reads as ShortRead objects}

\item{dir}{Save directory}

\item{filename1}{Name of file 1}

\item{filename2}{Name of file 2}
}
\value{
Named list of filepaths
}
\description{
Write reads to file
}
\keyword{internal}

