\name{writeFeatureCountsHTML}
\alias{writeFeatureCountsHTML}
\title{writeFeatureCountsHTML}
\usage{
writeFeatureCountsHTML(outfile, dirPath, ExonsCoveredTable,
  GenomicFeaturesTable, GenomicFeaturesDetectedTable)
}
\arguments{
  \item{outfile}{a path}

  \item{dirPath}{a path}

  \item{ExonsCoveredTable}{a table}

  \item{GenomicFeaturesTable}{a table}

  \item{GenomicFeaturesDetectedTable}{a table}
}
\value{
Nothing
}
\description{
writeFeatureCountsHTML
}
\author{
Gregoire Pau
}
\keyword{internal}

