% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPipelineQA.R
\name{writeGenomicFeaturesReport}
\alias{writeGenomicFeaturesReport}
\title{Generate pipeline report}
\usage{
writeGenomicFeaturesReport()
}
\value{
Name of created HTML file
}
\description{
Generates a summary HTML for the Genomic Feature counting step
}
\author{
Melanie Huntley, Cory Barr, Jens Reeder
}
\keyword{internal}

