% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaries.R
\name{writeSummary}
\alias{writeSummary}
\title{Write HTML summary}
\usage{
writeSummary(dirs, cutoffs, outdir = "./")
}
\arguments{
\item{dirs}{List of pipeline result dirs}

\item{cutoffs}{list, cutoffs for each plotting/QA function}

\item{outdir}{Path to output directory. Does not create dir.}
}
\value{
Nothing, but writes file
}
\description{
Write html Summary for list of runs
}
\author{
Jens Reeder
}
\keyword{internal}

