% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeVCF}
\alias{writeVCF}
\title{writeVCF}
\usage{
writeVCF(variants.vranges)
}
\arguments{
\item{variants.vranges}{Genomic Variants as VRanges object}
}
\value{
VCF file name
}
\description{
Write variants to VCF file
}
\author{
Jens Reeder
}

