% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{bamCountUniqueReads}
\alias{bamCountUniqueReads}
\title{Uniquely count number of reads in bam file}
\usage{
bamCountUniqueReads(bam)
}
\arguments{
\item{bam}{Name of bam file}
}
\value{
number of reads
}
\description{
Uniquely count number of reads in bam file
}
\author{
Jens Reeder
}
\keyword{internal}

