\name{buildConfig}
\alias{buildConfig}
\title{Build a configuration file based on a list of parameters}
\usage{
  buildConfig(config_filename, ...)
}
\arguments{
  \item{config_filename}{The path of a configuration
  filename.}

  \item{...}{A list of named value pairs.}
}
\value{
  Nothing.
}
\description{
  Build a configuration file based on a list of parameters
}
\author{
  Gregoire Pau
}
\seealso{
  runPipeline
}
\keyword{internal}

