% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildTP53GenomeTemplate}
\alias{buildTP53GenomeTemplate}
\title{buildTP53GenomeTemplate}
\usage{
buildTP53GenomeTemplate()
}
\value{
Path to tp53 template file
}
\description{
Create a tp53 config template
}
\author{
Jens Reeder
}
\keyword{internal}

