% Generated by roxygen2 (4.0.2): do not edit by hand
\name{countGenomicFeatures}
\alias{countGenomicFeatures}
\title{Count overlaps with genomic features}
\usage{
countGenomicFeatures()
}
\value{
Nothing
}
\description{
Count overlaps with genomic features
}
\author{
Gegoire Pau
}
\keyword{internal}

