% Generated by roxygen2 (4.0.2): do not edit by hand
\name{countGenomicFeaturesChunk}
\alias{countGenomicFeaturesChunk}
\title{Count reads by genomic Feature}
\usage{
countGenomicFeaturesChunk(save_dir)
}
\arguments{
\item{save_dir}{PAth to a pipeline run's save dir}
}
\value{
Nothing
}
\description{
Count reads by genomic Feature
}
\details{
given a BAM-file output from gsnap (with the MD tag), count
hits to exons, genes, ncRNAs, etc.
and quantify miRNA/ncRNA contaminatino
}
\author{
Cory Barr
}
\keyword{internal}

