% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{FastQStreamer.release}
\alias{FastQStreamer.release}
\title{Close the FastQStreamer}
\usage{
FastQStreamer.release()
}
\value{
Nothing
}
\description{
Close the FastQStreamer
}
\author{
Gregoire Pau
}
\seealso{
FastQStreamer.init
}
\keyword{internal}

