% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findVariantFile}
\alias{findVariantFile}
\title{Get a vcf filename given a HTSeqGenie directory}
\usage{
findVariantFile(save_dir)
}
\arguments{
\item{dir_path}{A character string containing a dir path}
}
\value{
A character vector containing an existing filename, stops if 0 or more than 1
}
\description{
Get the filename of the variant file
}
\details{
Depending on the variant caller used and the version of
VariantAnotation used to create the file a file might have the
ending vcf.gz, vcf.bgz. To function hides all this mess.
}
\author{
Jens Reeder
}
\keyword{internal}

