% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gatk.R
\name{gatk}
\alias{gatk}
\title{gatk}
\usage{
gatk(gatk.jar.path = getOption("gatk.path"), method, args, maxheap = "4g")
}
\arguments{
\item{gatk.jar.path}{Path to the gatk jar file}

\item{method}{Name of the gatk method, e.g. UnifiedGenotyper}

\item{args}{additional args passed to gatk}

\item{maxheap}{Maximal heap space allocated for java,
GATK recommends 4G heap for most of its apps}
}
\value{
0 for success, stops otherwise
}
\description{
Run a command from the GATK
}
\details{
Execute the GATK jar file using
the method specified as arg.
Stops if the command executed fails.
}
\author{
Jens Reeder
}

