% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\name{getConfig.numeric}
\alias{getConfig.numeric}
\title{Check if a config parameter is a numeric}
\usage{
getConfig.numeric(p, ...)
}
\arguments{
\item{p}{Name of parameter}

\item{...}{Extra params passed to getConfig}
}
\value{
Value of parameter as numeric
}
\description{
Throws exception if value can't be cast into numeric
}
\keyword{internal}

