\name{getNumberOfReadsInFASTQFile}
\alias{getNumberOfReadsInFASTQFile}
\title{Count reads in Fastq file}
\usage{
  getNumberOfReadsInFASTQFile(filename)
}
\arguments{
  \item{filename}{Name of FastQ file}
}
\value{
  Number of reads
}
\description{
  Count reads in Fastq file
}
\author{
  Gregoire Pau
}
\keyword{internal}

