\name{getNumericVectorDataFromFile}
\alias{getNumericVectorDataFromFile}
\title{Load data as numerical values}
\usage{
  getNumericVectorDataFromFile(dirPath, object_name)
}
\arguments{
  \item{dirPath}{Save dir of a pipeline run}

  \item{object_name}{Object name}
}
\value{
  loaded data as table of numbers
}
\description{
  Load data as numerical values
}
\author{
  Jens Reeder
}
\keyword{internal}

