% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectAdapterContam.R
\name{getRandomAlignCutoff}
\alias{getRandomAlignCutoff}
\title{Estimate an adapter alignment cutoff score}
\usage{
getRandomAlignCutoff(read_len, n)
}
\arguments{
\item{read_len}{The read length}

\item{n}{Number of samples}
}
\description{
Empirically estimate a threshold that discriminates
random reads from reads with adapter contamination
}
\keyword{internal}

