% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getTraceback}
\alias{getTraceback}
\title{Get traceback from tryKeepTraceback()}
\usage{
getTraceback(mto)
}
\arguments{
\item{mto}{An object of the try-error class}
}
\value{
Traceback as a string
}
\description{
Get traceback from tryKeepTraceback()
}
\keyword{internal}

