% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initPipelineFromConfig}
\alias{initPipelineFromConfig}
\title{Init pipeline environment}
\usage{
initPipelineFromConfig(config_filename, config_update)
}
\arguments{
\item{config_filename}{Name of config file}

\item{config_update}{List of name value pairs that will update the config parameters}
}
\value{
Nothing
}
\description{
Init pipeline environment
}
\author{
Jens Reeder
}
\keyword{internal}

