% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{listIterator.next}
\alias{listIterator.next}
\title{Get reads from the listIterator}
\usage{
listIterator.next()
}
\value{
An object. NULL if there are no more objects in the listIterator.
}
\description{
Get reads from the listIterator
}
\author{
Gregoire Pau
}
\seealso{
listIterator.init
}
\keyword{internal}

