% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logging.R
\name{logdebug}
\alias{logdebug}
\title{Log debug using the logging package}
\usage{
logdebug(msg)
}
\arguments{
\item{...}{Arguments passed to logging::logdebug}
}
\value{
Nothing
}
\description{
Log debug (with a try statement)
}
\author{
Gregoire Pau
}
\keyword{internal}

