% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logerror}
\alias{logerror}
\title{Log info using the logging package}
\usage{
logerror(msg)
}
\arguments{
\item{...}{Arguments passed to logging::loginfo}
}
\value{
Nothing
}
\description{
Log error (with a try statement)
}
\author{
Gregoire Pau
}
\keyword{internal}

