% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logwarn}
\alias{logwarn}
\title{Log warning using the logging package}
\usage{
logwarn(msg)
}
\arguments{
\item{...}{Arguments passed to logging::logwarn}
}
\value{
Nothing
}
\description{
Log warning (with a try statement)
}
\author{
Gregoire Pau
}
\keyword{internal}

