% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coverage.R
\name{mergeCoverage}
\alias{mergeCoverage}
\title{Merge coverage files}
\usage{
mergeCoverage(indirs, outdir, prepend_str)
}
\arguments{
\item{indirs}{A character vector, indicating which directories have to be merged}

\item{outdir}{A character string indicating the output directory (which must exist)}

\item{prepend_str}{A character string, containing a prefix going to be appended on all output result files}
}
\value{
Nothing
}
\description{
Merge coverage files
}
\details{
Merges coverage objects, usually SipleRleLists, in a tree-reduce
fashion. The coverage object dynamically switches to a SimpleIntegerList,
once the data becomes too dense.
}
\author{
Jens Reeder
}
\keyword{internal}

