% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDF}
\alias{plotDF}
\title{Make continuous plots of distribution function}
\usage{
plotDF(df, ylab, xlab, filename)
}
\arguments{
\item{df}{distribution function, given as absolute count and percent}

\item{ylab}{label of y axis}

\item{xlab}{label of x axis}

\item{filename}{plots will be saved under [filename].png and [filename].pdf}
}
\value{
Nothing, creates two files instead
}
\description{
Make continuous plots of distribution function
}
\author{
Jens Reeder
}
\keyword{internal}

