\name{runAlignment}
\alias{runAlignment}
\title{Runs the read alignment step of the pipeline}
\usage{
  runAlignment(config_filename, config_update)
}
\arguments{
  \item{config_filename}{Path to configuration file}

  \item{config_update}{List of name value pairs that will
  update the config parameters}
}
\value{
  Nothing
}
\description{
  Runs the read alignment step of the pipeline
}
\author{
  Jens Reeder
}
\keyword{internal}

