% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipeline.R
\name{runPipelineConfig}
\alias{runPipelineConfig}
\title{Run the NGS analysis pipeline}
\usage{
runPipelineConfig(config_filename, config_update)
}
\arguments{
\item{config_filename}{Path to a pipeline configuration file}

\item{config_update}{A list of name value pairs that will update the config parameters}
}
\value{
Nothing
}
\description{
Run the NGS analysis pipeline from a configuration file
}
\details{
This is the launcher function for all
pipeline runs. It will do some
preprocessing steps, then aligns the reads,
counts overlap with genomic Features such as genes, exons etc
and applies a variant caller.
}
\author{
Jens Reeder, Gregoire Pau
}

