% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{setupTestFramework}
\alias{setupTestFramework}
\title{setup test framework}
\usage{
setupTestFramework(config.filename, config.update = list(),
  testname = "test", package = "HTSeqGenie", use.TP53Genome = TRUE)
}
\arguments{
\item{config.filename}{configuration file}

\item{config.update}{update list of config values}

\item{testname}{name of test case}

\item{package}{name of package}

\item{use.TP53Genome}{Boolean indicating the use of the TP53 genome as template config}
}
\value{
the created temp directory
}
\description{
setup test framework
}

