% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimReads.R
\name{trimReads}
\alias{trimReads}
\title{Trim/truncate a set of reads}
\usage{
trimReads(lreads, trim_len = NULL, trim5 = 0)
}
\arguments{
\item{lreads}{A list of ShortReadQ objects}

\item{trim_len}{The length reads will be truncated to; default is NULL (no length truncation)}

\item{trim5}{The number of nucleotides to trim from the 5'-end; default is 0}
}
\value{
A list of truncated ShortReadQ objects
}
\description{
Trim/truncate a set of reads
}
\keyword{internal}

