% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trimReads.R
\name{truncateReads}
\alias{truncateReads}
\title{Trim/truncate a set of reads}
\usage{
truncateReads(reads, trim_len = NULL, trim5 = 0)
}
\arguments{
\item{reads}{A set of reads as ShortReadQ object}

\item{trim_len}{The length reads will be truncated to; default is NULL (no length truncation)}

\item{trim5}{The number of nucleotides to trim from the 5'-end; default is 0}
}
\value{
A truncated ShortReadQ object
}
\description{
Trim/truncate a set of reads
}
\keyword{internal}

