% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\name{updateConfig}
\alias{updateConfig}
\title{Update the existing config}
\usage{
updateConfig(tconfig)
}
\arguments{
\item{tconfig}{List of configuration name value pairs}
}
\value{
Nothing.
}
\description{
Update the existing config
}
\keyword{internal}

