\name{writeAudit}
\alias{writeAudit}
\title{Write Session information}
\usage{
  writeAudit(filename)
}
\arguments{
  \item{filename}{Optional name of file. If missing, prints
  session information on the standard output.}
}
\value{
  Nothing
}
\description{
  Write Session information
}
\author{
  Gregoire Pau
}
\keyword{internal}

