\name{writePreprocessAlignReport}
\alias{writePreprocessAlignReport}
\title{Generate Pipeline Report}
\usage{
writePreprocessAlignReport()
}
\value{
Name of created HTML file
}
\description{
Generates a summary HTML for the preprocess and align step
}
\author{
Melanie Huntley, Cory Barr, Jens Reeder
}
\keyword{internal}

